requirements.txt
pandas>=2.0.0
paho-mqtt>=1.6.0
matplotlib>=3.7.0
rich>=13.0.0